Change the summary list to a summary card
This also removes some of the notes-related fields, which will be
rendered in a different component in a future commit
