+++ lib/engines/content_block_manager/app/components/content_block_manager/content_block_edition/show/confirm_summary_card_component.html.erb
@@ -1,0 +1,5 @@
+<%= render "components/summary_card", {
+  title:,
+  rows:,
+  summary_card_actions:,
+} %>
