+++ lib/engines/content_block_manager/app/views/content_block_manager/content_block/editions/workflow/review.html.erb
@@ -17,7 +17,7 @@
 <div class="govuk-grid-row">
   <div class="govuk-grid-column-full">
     <%= render(
-          ContentBlockManager::ContentBlockEdition::Show::ConfirmSummaryListComponent.new(
+          ContentBlockManager::ContentBlockEdition::Show::ConfirmSummaryCardComponent.new(
             content_block_edition: @content_block_edition,
             ),
           ) %>
